SASS-3070 - [Refactoring] Renames and moves classes in relevant packages.
