+++ test/support/TaxYearUtils.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package utils
+package support
 
 import java.time.LocalDate
