+++ test/support/builders/models/pages/ContractorCYAPageBuilder.scala
@@ -17,7 +17,7 @@
 package support.builders.models.pages
 
 import models.pages.ContractorCYAPage
-import utils.TaxYearUtils
+import support.TaxYearUtils
 
 import java.time.Month
