+++ test/support/builders/models/pages/DeductionAmountPageBuilder.scala
@@ -18,7 +18,7 @@ package support.builders.models.pages
 
 import forms.FormsProvider
 import models.pages.DeductionAmountPage
-import utils.TaxYearUtils
+import support.TaxYearUtils
 
 import java.time.Month
