+++ test/support/builders/models/pages/DeleteCISPeriodPageBuilder.scala
@@ -17,7 +17,7 @@
 package support.builders.models.pages
 
 import models.pages.DeleteCISPeriodPage
-import utils.TaxYearUtils.taxYearEOY
+import support.TaxYearUtils.taxYearEOY
 
 import java.time.Month
