+++ test/support/builders/models/pages/LabourPayPageBuilder.scala
@@ -18,7 +18,7 @@ package support.builders.models.pages
 
 import forms.FormsProvider
 import models.pages.LabourPayPage
-import utils.TaxYearUtils
+import support.TaxYearUtils
 
 import java.time.Month
