+++ test/support/builders/models/pages/MaterialsAmountPageBuilder.scala
@@ -18,7 +18,7 @@ package support.builders.models.pages
 
 import forms.FormsProvider
 import models.pages.MaterialsAmountPage
-import utils.TaxYearUtils
+import support.TaxYearUtils
 
 import java.time.Month
